requirements.txt
# Core dependencies
python-telegram-bot==21.0.1
APScheduler==3.10.4
PyYAML==6.0.1

# Compatibility fixes
attrs>=23.2.0

# Database
psycopg[binary]==3.2.3

# ML and data processing
scikit-learn==1.5.2
numpy==2.1.3
pandas==2.2.3

# Web scraping
beautifulsoup4==4.12.3
requests==2.32.3

# Visualization
wordcloud==1.9.3
Pillow==11.0.0
matplotlib==3.9.2

# Utilities
python-dateutil==2.9.0
